requirements.txt
sphinx>=2.0
sphinx_rtd_theme
sphinx-autodoc-typehints
numpy
scipy
pandas<1.0.0
pyyaml>=5.1
schema
plams@git+https://github.com/SCM-NV/PLAMS@master
qmflows@git+https://github.com/SCM-NV/qmflows@master
nano-cat@git+https://github.com/nlesc-nano/nano-CAT@devel
data-cat@git+https://github.com/nlesc-nano/data-CAT@devel
